Factor out DragAndDropThought component.
